chore(suite-native): Mobile Trade: memoize TradeHistoryItem
